processing-service/requirements.txt
fastapi
uvicorn
PyPDF2
Pillow
python-dotenv